Bug 794510: Part 6 - Fixups needed to build on Windows and Android and Mac OS X debug, plus shorten comment to be less than 80 columns; r=ehsan
UltraBlame original commit: 62627192e1563789da8b97907ffd342f4348f31d

+++ tools/trace-malloc/lib/nsDebugHelpWin32.cpp
@@ -224,7 +224,7 @@ DHWImportHooker::PatchOneModule(HMODULE aModule, const char* name)
     
 
     PIMAGE_IMPORT_DESCRIPTOR desc;
-    uint32_t size;
+    ULONG size;
 
     desc = (PIMAGE_IMPORT_DESCRIPTOR) 
         dhwImageDirectoryEntryToData(aModule, true,
